refactoring

+++ db/migrations/20221207075911_create_table_product_addons.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS addons(
     id BIGSERIAL PRIMARY KEY NOT NULL,
     name VARCHAR(255),
     description VARCHAR(255),
-    price FLOAT,
+    price NUMERIC,
     created_at BIGINT NOT NULL DEFAULT extract(epoch from now()),
     updated_at BIGINT
 );
